src/fmp_py/StockAnalysis/database/requirements.txt
# FMP MySQL Cache Requirements
# Python dependencies for FMP database caching system

# Database connector
mysql-connector-python>=8.0.33

# Environment variables
python-dotenv>=1.0.0

# Data processing (already used by FMP client)
pandas>=2.0.0
numpy>=1.24.0

# HTTP client (already used by FMP client)
requests>=2.31.0

# Optional: For advanced database operations
# SQLAlchemy>=2.0.0  # Uncomment if you want ORM support
